// frontend/src/api/generated/model/prIssueComment.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */
import type { PrIssueCommentReactions } from './prIssueCommentReactions';
import type { GitHubUser } from './gitHubUser';

export interface PrIssueComment {
  body: string;
  created_at: string;
  html_url: string;
  /** @minimum 0 */
  id: number;
  reactions?: PrIssueCommentReactions;
  updated_at: string;
  user: GitHubUser;
}